+++ app/aboutus/page.tsx
@@ -1,11 +1,13 @@
+
 import Navbar from "@/components/Navbar";
 import Image from "next/image";
 import Footer from "@/components/Footer";
 import Banner from "@/components/subcomponents/banner";
+import Addbanner from "@/components/Addbanner";
 
 type Props = {};
 
-const page = (props: Props) => {
+const About = (props: Props) => {
   return (
     <div>
       <Navbar />
@@ -24,117 +26,17 @@ const page = (props: Props) => {
       {/* ------- */}
       <div className="flex flex-col justify-center align-middle p-2 md:p-20">
         <div className="flex flex-col p-10">
-          <p className="text-black font-bold text-xl">1.Quality Policy</p>
-          <p className="text-black font-sm text-wrap text-md p-3">
-            At UNAVAR FOOD INSPECTION AND CERTIFICATION PRIVATE LIMITED, we are
-            committed to providing Hygiene Rating Certification in an impartial,
-            independent, and value-added manner to our clients while meeting the
-            requirements of accreditation bodies. We ensure certification
-            services are delivered by competent personnel and continually
-            enhance their competency through training and evaluations. Our
-            commitment to continual improvement involves reviewing certification
-            activities and gathering strong feedback from clients, auditors,
-            certification staff, and other stakeholders. We guarantee
-            accessibility to our services for all clients, free from financial
-            or any other bias, and operate all procedures and guidelines in a
-            non-discriminatory manner. Upholding policies of impartiality and
-            confidentiality, we ensure communication of these policies from top
-            management to all staff, stakeholders, and clients. Additionally,
-            our auditors are informed that there is no special remuneration tied
-            to the number of audits conducted, mitigating any potential
-            influence on audit results.
-          </p>
-        </div>
-        <div className="flex flex-col p-10">
-          <p className="text-black font-bold text-xl">2.Impartiality Policy</p>
-          <p className="text-black font-md text-wrap p-3">
-            At UNAVAR FOOD INSPECTION AND CERTIFICATION PRIVATE LIMITED, we
-            uphold impartiality in our certification and inspection activities,
-            ensuring fairness among all Food Business Operators (FBOs).
-            Certification is not restricted by financial or other limiting
-            conditions, and the company takes full responsibility for
-            maintaining impartiality and preventing external pressures from
-            compromising this. Disclosure of reports and documentation is
-            limited to involved FBOs and relevant authorities for compliance
-            reporting or conflict resolution. Internal procedures are in place
-            to ensure objectivity and manage conflicts of interest, adhering to
-            strict corporate integrity standards. Personnel are not influenced
-            by financial or commercial pressures and actively address situations
-            that may affect impartiality. Ongoing identification of threats to
-            impartiality is conducted, including self-interest, self-review,
-            familiarity, intimidation, and financial threats. UNAVAR&LSQUO;s
-            management system is equipped with necessary safeguards to enforce
-            these policies. The company&LSQUO;s certification services do not
-            interfere with client&LSQUO;s trade or audit quality, irrespective
-            of business volume. Additionally, UNAVAR refrains from engaging in
-            activities such as product certification, supply, or catering sector
-            training to mitigate risks to impartiality. A certification and
-            inspection scheme committee ensures balanced representation of
-            stakeholders in developing certification and inspection schemes.
-          </p>
-        </div>
-        <div className="flex flex-col p-10">
-          <p className="text-black font-bold text-xl">
-            3.Confidentiality Policy
-          </p>
-          <p className="text-black font-md text-wrap p-3">
-            Employees and personnel of UNAVAR FOOD INSPECTION AND CERTIFICATION
-            PRIVATE LIMITED are bound by strict confidentiality obligations
-            regarding information obtained during their employment or
-            engagement. This policy outlines their duties to maintain
-            confidentiality and refrain from disclosing information except for
-            lawful purposes. All staff members must handle information received
-            or accessible to them with utmost confidentiality and utilize it
-            only for authorized purposes in compliance with the law. UNAVAR FOOD
-            INSPECTION AND CERTIFICATION PRIVATE LIMITED, including its
-            certification and inspection body team, employees, and contractors,
-            commits to safeguarding all information obtained during audit or
-            certification activities or any other dealings with organizations.
-            Such information shall not be disclosed to third parties without
-            express permission, except when legally mandated or when safety
-            concerns arise. Records are securely maintained and accessible only
-            to authorized personnel, with contractors limited to accessing
-            information relevant to their audit activities. Information may be
-            disclosed to organizations with a legitimate right to access it,
-            including Accreditation Bodies, ensuring compliance with legal
-            requirements and safeguarding confidentiality.
-          </p>
-        </div>
-        <div className="flex flex-col p-10">
-          <p className="text-black font-bold text-xl">4.THIRD PARTY AUDIT</p>
-          <p className="text-black font-md text-wrap p-3">
-            Who all are eligible for TPA? The Food Business Operators (FBOs) who
-            hold a central license and operate a manufacturing unit are eligible
-            to undergo Third Party Audit (TPA). In the context of food safety
-            regulations, TPA refers to an audit conducted by an external,
-            independent organization or body that is recognized by relevant
-            authorities. This audit assesses the compliance of the food business
-            with applicable standards and regulations, ensuring that food safety
-            and quality requirements are met. On What Basis Man - Days were
-            Calculated? Man-days refer to the number of days worked by an
-            individual. In the context of food handling, man-days are calculated
-            based on the number of food handlers employed by a food business.
-            The provided breakdown suggests different man- day allocations based
-            on the number of food handlers: 0 - 50 food handlers: 0.5 man-day 50
-            - 100 food handlers: 1 man-day More than 100 food handlers: 1.5
-            man-days This calculation is likely used for determining the
-            resources required for food safety management, including training,
-            supervision, and monitoring of food handling practices. What is TPA
-            and how it is useful? TPA, or Third Party Audit, is an evaluation
-            process carried out by an independent organization or body that is
-            officially recognized for its competence and authority in assessing
-            compliance with relevant standards and regulations. These recognized
-            bodies typically have accreditation from regulatory agencies or
-            industry associations. The purpose of TPA is to provide an impartial
-            assessment of a food business&LSQUO;s compliance with food safety
-            and quality standards, offering assurance to consumers and
-            regulatory authorities alike
+          <p>
+            Unavar Food Inspection and Certification Private Limited is an certification body which is formed to ensure the Food Business operators in India get Certified under different schemes of Food safety standards authority of India ( FSSAI) , standards of International Organization for Standardization ( ISO) and other Global Standards.
+            Unavar Food Inspection and Certification Private Limited&LSQUO;s vision is to make FBO&LSQUO;s undergo different certifications in any part of the country in ease manner to ensure the meet the requirements of National and Global Food Safety standards and provide safe and wholesome food to the consumers.
+            Unavar Food Inspection and Certification Private Limited aims to be connected with the society and Food business operators through young, talented and passionate professional who are trained and made competent to deliver knowledge about food safety to the industry. Unavar Food Inspection and Certification Private Limited provides certification to food business operators which ensures the consumers the following Safe Food is prepared and served Meets the requirements of National food safety standards or Global Standards Rating of the FBO according to the practices they follow
           </p>
         </div>
       </div>
       <footer>
         <div>
           <Banner />
+          <Addbanner />
           <Footer />
         </div>
       </footer>
@@ -142,4 +44,4 @@ const page = (props: Props) => {
   );
 };
 
-export default page;
+export default About;
